requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyjwt
passlib
bcrypt
psycopg2-binary